% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_blues.R
\name{sim_blue_model}
\alias{sim_blue_model}
\title{sim_blues}
\usage{
sim_blue_model(
  pars = list(I = 4, J = 3, Z = 3, R = 2, u_treat = 0, sigma_treat = 1, u_gen = 0,
    sigma_gen = 1, u_block = 0, sigma_block = 1, u_pb = 0, sigma_pb = 1, u_gp = 0,
    sigma_gp = 1, u_mean = 0, sigma_mean = 1, u_epsilon = 0, sigma_epsilon = 1)
)
}
\arguments{
\item{pars}{a list containing the number of genotypes I, environments J, Z, marginal mean of treatmens}
}
\description{
A function to simulate blues in a block design context.
}
