% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_m1.R
\name{sim_m1}
\alias{sim_m1}
\title{sim_m1}
\usage{
sim_m1(
  pars = list(w_a = w_a, u_mean = 0, sigma_mean = 1, sigma_ag = 1, u_env = 0, sigma_env =
    1, u_epsilon = 0, sigma_epsilon = 1, J = 3, I = 2, u_gen = 1)
)
}
\arguments{
\item{pars}{a list containing the parameter required to simulate from model 1}
}
\description{
A function to simulate from Model one proposed in
}
