% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_m2.R
\name{sim_m2}
\alias{sim_m2}
\title{sim_m2}
\usage{
sim_m2(
  pars = list(I = 2, J = 3, w_a = w_a, sigma_ag = 1, u_mean = 0, sigma_mean = 1, u_env =
    0, sigma_env = 1, u_gen = 0, sigma_env = 1, sigma_ge = 2, u_ge = 0, sigma_ge = 1,
    u_epsilon = 0, sigma_epsilon = 1)
)
}
\arguments{
\item{pars}{}
}
\description{
sim_m2
}
