% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_m4.R
\name{sim_m4}
\alias{sim_m4}
\title{sim_m4}
\usage{
sim_m4(
  pars = list(I = 2, J = 3, K = 10, w_a = w_a, sigma_a = 1, u_mean = 0, sigma_mean = 1,
    u_env = 0, sigma_env = 1, u_gen = 0, sigma_env = 1, u_epsilon = 0, sigma_epsilon = 1,
    x_env = x_env, sigma_w = 1, u_w = 0, sigma_ge = 1, u_ge = 0)
)
}
\arguments{
\item{pars}{}
}
\description{
Simulates from model 4
}
